% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mmm_project.R
\name{create_mmm_project}
\alias{create_mmm_project}
\title{Creat an MMM Project}
\usage{
create_mmm_project(path = "~/Desktop/", FiscalYear = NULL,
  modeling_start_date = NULL, modeling_end_date = NULL,
  add_drake_workflow = TRUE, open_proj = TRUE, init_packrat = FALSE)
}
\arguments{
\item{path}{- path to location where project folder should be created.}

\item{FiscalYear}{- string specifying the fiscal period for the project.}

\item{modeling_start_date}{- yyyy-mm-dd fromat string.}

\item{modeling_end_date}{- yyyy-mm-dd fromat string.}

\item{add_drake_workflow}{- logical. should a drake workflow subdirectory be created.}

\item{open_proj}{- logical. should the Rstudio project be opened when created.}

\item{init_packrat}{- logical. should a packrat folder be initiated automatically.}
}
\description{
Creat an MMM Project
}
