% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frammme-package.R
\docType{package}
\name{frammme}
\alias{frammme}
\alias{frammme-package}
\title{frammme: Framework for MMM projects}
\description{
Functions to Generate an MMM Project Framework.
The main function of this package is `create_mmm_project`
which takes parameters as arguments and creates a folder with an RStudio project following the MMM modeling framework.
}
\section{Functions}{

}

